+++ typescript/agentkit/CHANGELOG.md
@@ -8,6 +8,10 @@
 - Added `token_prices_by_symbol` action to fetch token prices by symbol.
 - Added `token_prices_by_address` action to fetch token prices by network and address pairs.
 
+### Fixed
+
+- Added account argument in call to estimateGas in CdpWalletProvider
+
 ## [0.1.1] - 2025-02-02
 
 ### Added